Supportive cat now gives % buff
Changed the buff type to be a percentile increase.

+++ Assets/Scripts/UnitScripts/CardScript.cs
@@ -42,8 +42,8 @@ public class CardScript : ScriptableObject
     [Tooltip("What layer is allies on?")] public LayerMask allyLayerToTarget;
     [Tooltip("The range that the support can reach")] public float hitAllyRange;
     [Tooltip("How much bigger will the unit be?")] public Vector3 sizeBuff;
-    [Tooltip("How much the health should be buffed")] public int healthBuff;
-    [Tooltip("How much the damage should be buffed")] public int damageBuff;
+    [Tooltip("How much the health should be buffed")] public float healthBuff;
+    [Tooltip("How much the damage should be buffed")] public float damageBuff;
 
     [Header("Test options, must be off to work normaly")]
     public bool isNotInCorrectSceneTest; //Will be removed
